Moved enums to package com.projectg2.thebookclub.enums
